+++ android/src/main/java/com/mr/flutter/plugin/filepicker/FilePickerDelegate.java
@@ -15,10 +15,13 @@ import android.os.Parcelable;
 import android.provider.DocumentsContract;
 import android.util.Log;
 
+import androidx.annotation.RequiresApi;
 import androidx.annotation.VisibleForTesting;
 import androidx.core.app.ActivityCompat;
 
 import java.io.File;
+import java.io.IOException;
+import java.io.OutputStream;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Objects;
@@ -31,6 +34,7 @@ public class FilePickerDelegate implements PluginRegistry.ActivityResultListener
 
     private static final String TAG = "FilePickerDelegate";
     private static final int REQUEST_CODE = (FilePickerPlugin.class.hashCode() + 43) & 0x0000ffff;
+    private static final int SAVE_FILE_CODE = (FilePickerPlugin.class.hashCode() + 83) & 0x0000ffff;
 
     private final Activity activity;
     private final PermissionManager permissionManager;
@@ -41,6 +45,8 @@ public class FilePickerDelegate implements PluginRegistry.ActivityResultListener
     private int compressionQuality=20;
     private String[] allowedExtensions;
     private EventChannel.EventSink eventSink;
+
+    private byte[] bytes;
 
     public FilePickerDelegate(final Activity activity) {
         this(
@@ -76,6 +82,38 @@ public class FilePickerDelegate implements PluginRegistry.ActivityResultListener
 
     @Override
     public boolean onActivityResult(final int requestCode, final int resultCode, final Intent data) {
+        // Save file
+        if (requestCode == SAVE_FILE_CODE) {
+            if (resultCode == Activity.RESULT_OK) {
+                this.dispatchEventStatus(true);
+                final Uri uri = data.getData();
+                if (uri != null) {
+                  String  path = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS)
+                            .getAbsolutePath() + File.separator + FileUtils.getFileName(uri, this.activity);
+                    try {
+                        OutputStream outputStream = this.activity.getContentResolver().openOutputStream(uri);
+                        if(outputStream != null){
+                            outputStream.write(bytes);
+                            outputStream.flush();
+                            outputStream.close();
+                        }
+                        finishWithSuccess(path);
+                        return true;
+                    } catch (IOException e) {
+                        Log.i(TAG, "Error while saving file", e);
+                        finishWithError("Error while saving file", e.getMessage());
+                    }
+                }
+
+            }
+            if (resultCode == Activity.RESULT_CANCELED) {
+                Log.i(TAG, "User cancelled the save request");
+                finishWithSuccess(null);
+            }
+            return false;
+        }
+
+        // Pick files
         if (type == null) {
             return false;
         }
@@ -288,6 +326,39 @@ public class FilePickerDelegate implements PluginRegistry.ActivityResultListener
             }
         }
         this.startFileExplorer();
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.KITKAT)
+    public void saveFile(String fileName, String type, String initialDirectory, String[] allowedExtensions, byte[] bytes, MethodChannel.Result result) {
+        if (!this.setPendingMethodCallAndResult(result)) {
+            finishWithAlreadyActiveError(result);
+            return;
+        }
+        Intent intent = new Intent(Intent.ACTION_CREATE_DOCUMENT);
+        intent.addCategory(Intent.CATEGORY_OPENABLE);
+        if (fileName != null && !fileName.isEmpty()) {
+            intent.putExtra(Intent.EXTRA_TITLE, fileName);
+        }
+        this.bytes = bytes;
+        if (type != null && !"dir".equals(type) && type.split(",").length == 1) {
+            intent.setType(type);
+        } else {
+            intent.setType("*/*");
+        }
+        if (initialDirectory != null && !initialDirectory.isEmpty()) {
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                intent.putExtra(DocumentsContract.EXTRA_INITIAL_URI, Uri.parse(initialDirectory));
+            }
+        }
+        if (allowedExtensions != null && allowedExtensions.length > 0) {
+            intent.putExtra(Intent.EXTRA_MIME_TYPES, allowedExtensions);
+        }
+        if (intent.resolveActivity(this.activity.getPackageManager()) != null) {
+            this.activity.startActivityForResult(intent, SAVE_FILE_CODE);
+        } else {
+            Log.e(TAG, "Can't find a valid activity to handle the request. Make sure you've a file explorer installed.");
+            finishWithError("invalid_format_type", "Can't handle the provided file type.");
+        }
     }
 
     @SuppressWarnings("unchecked")