+++ android/src/main/java/com/mr/flutter/plugin/filepicker/FilePickerPlugin.java
@@ -156,6 +156,16 @@ public class FilePickerPlugin implements MethodChannel.MethodCallHandler, Flutte
             return;
         }
 
+        if (call.method != null && call.method.equals("save")) {
+            String fileName = (String) arguments.get("fileName");
+            String type = resolveType((String) arguments.get("fileType"));
+            String initialDirectory = (String) arguments.get("initialDirectory");
+            String[] allowedExtensions = FileUtils.getMimeTypes((ArrayList<String>) arguments.get("allowedExtensions"));
+            byte[] bytes = (byte[]) arguments.get("bytes");
+            this.delegate.saveFile(fileName, type, initialDirectory, allowedExtensions, bytes,result);
+            return;
+        }
+
         fileType = FilePickerPlugin.resolveType(call.method);
         String[] allowedExtensions = null;