+++ lib/src/windows/file_picker_windows.dart
@@ -166,6 +166,7 @@ class FilePickerWindows extends FilePicker {
     String? initialDirectory,
     FileType type = FileType.any,
     List<String>? allowedExtensions,
+    Uint8List? bytes,
     bool lockParentWindow = false,
   }) async {
     final port = ReceivePort();